+++ src/components/SEO.jsx
@@ -1,76 +1,75 @@
 import React from 'react'
+import { Helmet } from 'react-helmet'
 import PropTypes from 'prop-types'
-import Helmet from 'react-helmet'
 import { StaticQuery, graphql } from 'gatsby'
-
-function SEO({ description, lang, meta, keywords, title }) {
-  return (
-    <StaticQuery
-      query={detailsQuery}
-      render={data => {
-        const metaDescription = description || data.site.siteMetadata.description
-        return (
-          <Helmet
-            htmlAttributes={{ lang: lang || 'en' }}
-            title={title || data.site.siteMetadata.title}
-            titleTemplate={`%s | ${ data.site.siteMetadata.title }`}
-            meta={[
-              {
-                name: 'description',
-                content: metaDescription,
-              },
-              {
-                property: 'og:title',
-                content: title,
-              },
-              {
-                property: 'og:description',
-                content: metaDescription,
-              },
-              {
-                property: 'og:type',
-                content: 'website',
-              },
-            ]
-              .concat(
-                keywords.length > 0
-                  ? {
-                    name: 'keywords',
-                    content: keywords.join(', '),
-                  }
-                  : []
-              )
-              .concat(meta)}
-          />
-        )
-      }}
-    />
-  )
-}
-
-SEO.defaultProps = {
-  lang: 'en',
-  meta: [],
-  keywords: [],
-}
-
-SEO.propTypes = {
-  description: PropTypes.string,
-  lang: PropTypes.string,
-  meta: PropTypes.array,
-  keywords: PropTypes.arrayOf(PropTypes.string),
-  title: PropTypes.string.isRequired,
-}
-
+const SEO = ({ title, description, image, pathname, article }) => (
+  <StaticQuery
+    query={query}
+    render={({
+      site: {
+        siteMetadata: {
+          defaultTitle,
+          defaultDescription,
+          siteUrl,
+          defaultImage,
+        },
+      },
+    }) => {
+      const seo = {
+        title: title || defaultTitle,
+        description: description || defaultDescription,
+        image: `${ siteUrl }${ image || defaultImage }`,
+        url: `${ siteUrl }${ pathname || '/' }`,
+      }
+      return (
+        <>
+          <Helmet title={seo.title}>
+            <meta name="description" content={seo.description} />
+            <meta name="image" content={seo.image} />
+            {seo.url && <meta property="og:url" content={seo.url} />}
+            {(article ? true : null) && (
+              <meta property="og:type" content="article" />
+            )}
+            {seo.title && <meta property="og:title" content={seo.title} />}
+            {seo.description && (
+              <meta property="og:description" content={seo.description} />
+            )}
+            {seo.image && <meta property="og:image" content={seo.image} />}
+            <meta name="twitter:card" content="summary_large_image" />
+            {seo.description && (
+              <meta name="twitter:description" content={seo.description} />
+            )}
+            {seo.image && <meta name="twitter:image" content={seo.image} />}
+            <html lang="en" />
+          </Helmet>
+        </>
+      )
+    }}
+  />
+)
 export default SEO
-
-const detailsQuery = graphql`
-  query DefaultSEOQuery {
+SEO.propTypes = {
+  title: PropTypes.string,
+  description: PropTypes.string,
+  image: PropTypes.string,
+  pathname: PropTypes.string,
+  article: PropTypes.bool,
+}
+SEO.defaultProps = {
+  title: null,
+  description: null,
+  image: null,
+  pathname: null,
+  article: false,
+}
+const query = graphql`
+  query SEO {
     site {
       siteMetadata {
-        title
-        description
-        author
+        defaultTitle: title
+        defaultDescription: description
+        siteUrl: url
+        defaultImage: image
       }
     }
   }