+++ src/components/elements/ElementUsage.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { StaticQuery, graphql } from 'gatsby'
 import withStyles from 'react-jss'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
-import classnames from 'classnames'
+import clsx from 'clsx'
 
 const styles = theme => ({
   root: {
@@ -41,11 +41,10 @@ const ElementUsage = ({ element, classes }) => {
         const documentation = data.allMdx.edges.find(
           x => x.node.frontmatter.label === `${ element }Usage`
         )
-
         return documentation ? (
           <MDXRenderer>{documentation.node.body}</MDXRenderer>
         ) : (
-          <p className={classnames(classes.text, 'zep-typo--normal-3')}>
+          <p className={clsx(classes.text, 'zep-typo--normal-body1')}>
             No content yet, please check back later
           </p>
         )