+++ src/components/table/ContentTable.jsx
@@ -1,12 +1,14 @@
-import React from 'react'
+import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import withStyles from 'react-jss'
-import classnames from 'classnames'
+import clsx from 'clsx'
 import Table from './Table'
 import TableHeader from './TableHeader'
 import TableCell from './TableCell'
 import TableBody from './TableBody'
 import TableRow from './TableRow'
+import Tabs from '../tabs/Tabs'
+import Tab from '../tabs/Tab'
 
 const styles = theme => ({
   root: {
@@ -21,83 +23,136 @@ const styles = theme => ({
   capitalize: {
     textTransform: 'capitalize',
   },
+  tabs: {
+    marginBottom: `${ theme.spacing.component.l.rem }rem`,
+  },
+  inverse: {
+    background: theme.colors.gray.black.hex,
+    color: 'rgba(255, 255, 255, 0.84)',
+  },
+  inverseLight: {
+    background: theme.colors.gray.grayMidDark.hex,
+    color: 'rgba(255, 255, 255, 0.84)',
+  },
 })
 
 const ContentTable = ({ header, content, title, classes }) => {
+  const [value, setValue] = useState(0)
+
+  function handleChange(newValue) {
+    setValue(newValue)
+  }
   return (
-    <div className={classes.root}>
-      <Table className={classes.table} title={title}>
-        <TableHeader>
-          {header.map(elem => (
-            <TableCell
-              className={classnames({
-                [classes.lowercase]: elem === 'rem' || elem === 'px',
-                [classes.capitalize]: elem !== 'rem' && elem !== 'px',
-              })}
-              key={`header_${ elem }`}
-            >
-              {elem}
-            </TableCell>
-          ))}
-        </TableHeader>
-        {content && (
-          <TableBody>
-            {Object.keys(content).map((elem, idx) => (
-              <TableRow key={`row_${ elem }`} color={idx % 2 === 0 ? 'white' : 'gray'}>
-                {header.map((item, idx) =>
-                  idx === 0 ? (
-                    <TableCell key={`typography${ idx }`} scope="row" component="th">
-                      {item === 'actual size' ? (
-                        'lorem ipsum'
-                      ) : item === 'name' ? (
-                        elem
-                      ) : item === 'specifications' ? (
-                        <span>
-                          {content[elem][item].map(p => (
-                            <p key={`specification_${ p }`} className="zep-typo--normal-2">
-                              {p}
-                            </p>
-                          ))}
-                        </span>
-                      ) : content[elem][item] ? (
-                        content[elem][item]
-                      ) : (
-                        '-'
-                      )}
-                    </TableCell>
-                  ) : (
-                    <TableCell
-                      className={classnames({
-                        [elem]: item === 'actual size',
-                      })}
-                      key={`typography${ idx }`}
-                    >
-                      {item === 'actual size' ? (
-                        'lorem ipsum'
-                      ) : item === 'name' ? (
-                        elem
-                      ) : item === 'specifications' ? (
-                        <span>
-                          {content[elem][item].map(p => (
-                            <p key={`specification_${ p }`} className="zep-typo--normal-2">
-                              {p}
-                            </p>
-                          ))}
-                        </span>
-                      ) : content[elem][item] ? (
-                        content[elem][item]
-                      ) : (
-                        '-'
-                      )}
-                    </TableCell>
-                  )
-                )}
-              </TableRow>
+    <>
+      <Tabs value={value} onClick={handleChange} className={classes.tabs}>
+        <Tab label="Positive" />
+        <Tab label="Negative" />
+      </Tabs>
+      <div className={classes.root}>
+        <Table className={classes.table} title={title}>
+          <TableHeader>
+            {header.map(elem => (
+              <TableCell
+                className={clsx({
+                  [classes.lowercase]: elem === 'rem' || elem === 'px',
+                  [classes.capitalize]: elem !== 'rem' && elem !== 'px',
+                  [classes.inverse]: value === 1,
+                })}
+                key={`header_${ elem }`}
+              >
+                {elem}
+              </TableCell>
             ))}
-          </TableBody>
-        )}
-      </Table>
-    </div>
+          </TableHeader>
+          {content && (
+            <TableBody>
+              {Object.keys(content).map((elem, idx) => (
+                <TableRow
+                  key={`row_${ elem }`}
+                  color={idx % 2 === 0 ? 'white' : 'gray'}
+                  className={
+                    value === 1
+                      ? idx % 2 === 0
+                        ? classes.inverseLight
+                        : classes.inverse
+                      : null
+                  }
+                >
+                  {header.map((item, idx) =>
+                    idx === 0 ? (
+                      <TableCell
+                        key={`typography${ idx }`}
+                        scope="row"
+                        component="th"
+                      >
+                        {item === 'actual size' ? (
+                          'lorem ipsum'
+                        ) : item === 'name' ? (
+                          elem
+                        ) : item === 'specifications' ? (
+                          <span>
+                            {Object.keys(content[elem][item]).map(p => (
+                              <p
+                                key={`specification_${ p }`}
+                                className={
+                                  value === 1
+                                    ? 'zep-typo--normal-body2-negative'
+                                    : 'zep-typo--normal-body2'
+                                }
+                              >
+                                {p}: {content[elem][item][p]}
+                              </p>
+                            ))}
+                          </span>
+                        ) : content[elem][item] ? (
+                          content[elem][item]
+                        ) : (
+                          '-'
+                        )}
+                      </TableCell>
+                    ) : (
+                      <TableCell
+                        className={clsx({
+                          [elem]: item === 'actual size' && value === 0,
+                          [`${ elem }-negative`]:
+                            item === 'actual size' && value === 1,
+                        })}
+                        key={`typography${ idx }`}
+                      >
+                        {item === 'actual size' ? (
+                          'lorem ipsum'
+                        ) : item === 'name' ? (
+                          elem
+                        ) : item === 'specifications' ? (
+                          <span>
+                            {Object.keys(content[elem][item]).map(p => (
+                              <p
+                                key={`specification_${ p }`}
+                                className={
+                                  value === 1
+                                    ? 'zep-typo--normal-body2-negative'
+                                    : 'zep-typo--normal-body2'
+                                }
+                              >
+                                {p}: {content[elem][item][p]}
+                              </p>
+                            ))}
+                          </span>
+                        ) : content[elem][item] ? (
+                          content[elem][item]
+                        ) : (
+                          '-'
+                        )}
+                      </TableCell>
+                    )
+                  )}
+                </TableRow>
+              ))}
+            </TableBody>
+          )}
+        </Table>
+      </div>
+    </>
   )
 }