+++ src/components/table/TableHeader.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import classnames from 'classnames'
+import clsx from 'clsx'
 
 function TableHeader({ children: childrenProp, className: classNameProp }) {
   const children = React.Children.map(childrenProp, child => {
@@ -8,14 +8,17 @@ function TableHeader({ children: childrenProp, className: classNameProp }) {
       return null
     }
     return React.cloneElement(child, {
-      className: classnames('zep-table__cell--header', child.props.className),
+      className: clsx('zep-table__cell--header', child.props.className),
       scope: 'col',
       component: 'th',
     })
   })
+
   return (
     <thead>
-      <tr className={classnames('zep-table__row zep-table__row--header', classNameProp)}>
+      <tr
+        className={clsx('zep-table__row zep-table__row--header', classNameProp)}
+      >
         {children}
       </tr>
     </thead>
@@ -23,7 +26,11 @@ function TableHeader({ children: childrenProp, className: classNameProp }) {
 }
 
 TableHeader.propTypes = {
-  children: PropTypes.oneOfType([PropTypes.string, PropTypes.array, PropTypes.object]),
+  children: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.array,
+    PropTypes.object,
+  ]),
   className: PropTypes.string,
 }